Move config handling into its own class
